refactor: plugin/utils.py 2,220줄을 책임별 모듈로 분할
- plugin/files.py (456줄): 플러그인 파일 저장/검증/압축 해제
- plugin/metadata.py (657줄): config/schema 메타데이터 파싱
- plugin/runtime.py (727줄): 컨테이너 실행 인자 조립 등 실행 준비
- plugin/builder.py (515줄): Docker 빌드/태깅 통합
- plugin/utils.py는 63줄 하위호환 facade로 잔류 (기존 import·테스트
  patch 경로 보존, 함수 시그니처·동작 불변)
- 함수 내부 지연 import 제거, builder↔worker 순환은 facade가 builder
  심볼을 재수출하지 않도록 경계 조정으로 해소
